Add files via upload
This is a part of the AniMix tools for customizing Blender for a better animation experience. 
You can download and help support the development and future of hybrid animation buy:  AniMixi here AniMix https://www.blendermarket.com/products/animix
Join the AniMix discord community here:  https://discord.gg/TpxMcpXVYX

 
The add-on exposes the fullscreen button in the Main Windows Topbar making for a more immersive and focused experience using Blender.